+++ capa/rules.py
@@ -54,7 +54,7 @@ META_KEYS = (
     "maec/malware-family",
     "maec/malware-category",
     "maec/malware-category-ov",
-    "author",
+    "authors",
     "description",
     "lib",
     "scope",
@@ -101,6 +101,9 @@ SUPPORTED_FEATURES: Dict[str, Set] = {
         capa.features.common.Characteristic("embedded pe"),
         capa.features.common.String,
         capa.features.common.Format,
+        capa.features.common.Class,
+        capa.features.common.Namespace,
+        capa.features.common.Characteristic("mixed mode"),
     },
     FUNCTION_SCOPE: {
         capa.features.common.MatchedRule,
@@ -134,6 +137,9 @@ SUPPORTED_FEATURES: Dict[str, Set] = {
         capa.features.common.Characteristic("indirect call"),
         capa.features.common.Characteristic("call $+5"),
         capa.features.common.Characteristic("cross section flow"),
+        capa.features.common.Characteristic("unmanaged call"),
+        capa.features.common.Class,
+        capa.features.common.Namespace,
     },
 }
 
@@ -259,6 +265,13 @@ def parse_feature(key: str):
         return capa.features.insn.Number
     elif key == "offset":
         return capa.features.insn.Offset
+    # TODO remove x32/x64 flavor keys once fixed master/rules
+    elif key.startswith("number/"):
+        logger.warning("x32/x64 flavor currently not supported and deprecated")
+        return capa.features.insn.Number
+    elif key.startswith("offset/"):
+        logger.warning("x32/x64 flavor currently not supported and deprecated")
+        return capa.features.insn.Offset
     elif key == "mnemonic":
         return capa.features.insn.Mnemonic
     elif key == "basic blocks":
@@ -280,8 +293,11 @@ def parse_feature(key: str):
     elif key == "format":
         return capa.features.common.Format
     elif key == "arch":
-
         return capa.features.common.Arch
+    elif key == "class":
+        return capa.features.common.Class
+    elif key == "namespace":
+        return capa.features.common.Namespace
     else:
         raise InvalidRule("unexpected statement: %s" % key)
 
@@ -1282,6 +1298,12 @@ class RuleSet:
                     logger.debug('using rule "%s" and dependencies, found tag in meta.%s: %s', rule.name, k, v)
                     rules_filtered.update(set(capa.rules.get_rules_and_dependencies(rules, rule.name)))
                     break
+                if isinstance(v, list):
+                    for vv in v:
+                        if tag in vv:
+                            logger.debug('using rule "%s" and dependencies, found tag in meta.%s: %s', rule.name, k, vv)
+                            rules_filtered.update(set(capa.rules.get_rules_and_dependencies(rules, rule.name)))
+                            break
         return RuleSet(list(rules_filtered))
 
     def match(self, scope: Scope, features: FeatureSet, va: int) -> Tuple[FeatureSet, ceng.MatchResults]: